Ocr/requirements.txt
opencv-python
numpy
Pillow
pytesseract
pymupdf
requests
python-Levenshtein
matplotlib
os
